+++ src/main/java/controller/PaymentController.java
@@ -1,5 +1,6 @@
 package controller;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -17,7 +18,9 @@ import org.springframework.http.MediaType;
 
 import models.Loan;
 import models.Payment;
+import models.PaymentPost;
 import service.PaymentService;
+import service.LoanService;
 
 @Controller
 @RestController
@@ -25,6 +28,9 @@ import service.PaymentService;
 public class PaymentController {
 	@Autowired
 	private PaymentService paymentService;
+	
+	@Autowired
+	private LoanService loanService;
 	
 	protected PaymentController() {}
 	
@@ -45,16 +51,57 @@ public class PaymentController {
 		}
 	}
 	
+	@RequestMapping(value = "/getwithloan/{paymentnumber}")
+	@GetMapping
+	public @ResponseBody PaymentPost getPaymentWithLoan(@PathVariable("paymentnumber") Long paymentnumber) {
+		Optional<Payment> paymentOptional = this.paymentService.getPayment(paymentnumber);
+		Payment payment;
+		if (paymentOptional.isPresent()) {
+			payment = paymentOptional.get();
+		} else {
+			return null;
+		}
+		PaymentPost paymentPost = new PaymentPost();
+		paymentPost.setLoannumber(payment.getLoan().getLoannumber());
+		paymentPost.setLoan(payment.getLoan());
+		paymentPost.setPaymentamount(payment.getPaymentamount());
+		paymentPost.setPaymentnumber(payment.getPaymentnumber());
+		return paymentPost;
+	}
+	
 	@RequestMapping(value = "/update/new", consumes = MediaType.APPLICATION_FORM_URLENCODED_VALUE)
 	@PostMapping
-	public @ResponseBody Boolean addNewBranch(Payment payment) {
+	public @ResponseBody Boolean addNewPayment(PaymentPost paymentPost) {
+		System.out.println("Trying to save payment: " + paymentPost.toCustomString());
+		Payment payment = new Payment();
+		payment.setPaymentamount(payment.getPaymentamount());
+		payment.setPaymentdate(LocalDateTime.now());
+		Optional<Loan> loanOptional = this.loanService.getLoan(paymentPost.getLoannumber());
+		Loan loan;
+		if (!loanOptional.isPresent()) {
+			return false;
+		}
+		loan = loanOptional.get();
+		loan.getPayments().add(payment);
+		payment.setLoan(loanOptional.get());
+		this.paymentService.savePayment(payment);
+		return this.loanService.saveLoan(loan);
+	}
+	
+	/*@RequestMapping(value = "/update/existing", consumes = MediaType.APPLICATION_FORM_URLENCODED_VALUE)
+	@PostMapping
+	public @ResponseBody Boolean updatePayment(PaymentPost payment) {
 		System.out.println("Trying to save payment: " + payment.toCustomString());
 		return this.paymentService.savePayment(payment);
-	}
+	}*/
 	
 	@RequestMapping(value = "/delete/{paymentnumber}")
 	@GetMapping
 	public @ResponseBody Boolean deletePayment(@PathVariable("paymentnumber") Long paymentnumber) {
+		Optional<Payment> paymentOptional = this.paymentService.getPayment(paymentnumber);
+		if (!paymentOptional.isPresent()) {
+			return false;
+		}
 		return this.paymentService.deletePayment(paymentnumber);
 	}
